+++ engine/bonus-malus/development_initiative.hpp
@@ -7,13 +7,13 @@ namespace engine {
 // DEVELOPMENT & INITIATIVE
 // ===================================================
 
-inline static constexpr int64_t INIT_BONUS_MG = 6;
+inline static constexpr int32_t INIT_BONUS_MG = 6;
 
-inline static constexpr int64_t INIT_BONUS_EG = 3;
+inline static constexpr int32_t INIT_BONUS_EG = 3;
 
-inline static constexpr int64_t EARLY_ROOK_PENALTY = -30;
+inline static constexpr int32_t EARLY_ROOK_PENALTY = -30;
 
-inline static constexpr int64_t DEVELOPMENT_BONUS = 10;
+inline static constexpr int32_t DEVELOPMENT_BONUS = 10;
 
 } // namespace engine
