+++ engine/bonus-malus/move_ordering.hpp
@@ -6,9 +6,9 @@ namespace engine {
 // ===================================================
 // MOVE ORDERING (SEARCH)
 // ===================================================
-static constexpr int64_t CHECK_BONUS = 50;
-static constexpr int64_t KILLER1_BONUS = 2000;
-static constexpr int64_t KILLER2_BONUS = 1900;
+static constexpr int32_t CHECK_BONUS = 50;
+static constexpr int32_t KILLER1_BONUS = 2000;
+static constexpr int32_t KILLER2_BONUS = 1900;
 
 } // namespace engine
