+++ engine/bonus-malus/queen.hpp
@@ -8,7 +8,7 @@ namespace engine {
 // ===================================================
 
 // reduced from -30
-inline static constexpr int64_t ATTACKED_QUEEN_PENALTY = -25;
+inline static constexpr int32_t ATTACKED_QUEEN_PENALTY = -25;
 
 } // namespace engine
